Porting menuitems for markdown-preview to 3.10+
This should fix jpfleury/gedit-markdown#3

Since 3.10 menu items can't be added via UIManager anymore (deprecated).
The *new* way is to have a seperate class of type `Gedit.AppActivatable`
that uses its own method `self.extend_menu(extension_point)` to get an
instance of `Gedit.MenuExtension` and extend the menu via this object.

The actions still have to be added to the window instance in the
original
`Gedit.WindowActivatable` class. They can be added directly to the
window with `window.add_action(action)`.

The actions have to be of type `Gio.Action`. For now I used two
`Gio.SimpleAction`
intead of action groups, but the old action groups are defined for
compatibility
with the rest of the code.

I'm not sure if action groups are necessary, if so the window also has a
method `window.insert_action_group(Gio.ActionGroup)` that could be used.
But the implementation of `Gio` action groups is different to the `Gtk`
action groups.